requirements.txt
streamlit
google-genai
pdfplumber
edge-tts